Lab 8
Don't know if any of this works as intended. But if you run only
Test(Server side) you can Add a json string and send to site, print the
employee list, or clear from console. I believe that is what we needed
to do. For that side. we just need to finish the UI and should be able
to hook it up to send the data out